Method Unit::ModSpellCastTime .
Facts:

It is used both for modify the cast time of all spells and the duration time of channelled spells
It returns and does nothing when a duration of a channelled spell should not be affected by haste
Some rare channelled spells has both cast time and channelling duration time (example: Seduction)
Problems:

When we try to modify the cast time of a channelled spell using that method, we can't do nothing if the duration shouldn't be affected by haste (but this should be unreleated with the cast time)
There are useless repeated checks
Solution:

Split this method into Unit::ModSpellCastTime and Unit::ModSpellDurationTime. The first should be used to modify the cast time only, and the latter should be used to modify the duration time only.
No need to check if spell duration should be affected by haste before calling Unit::ModSpellDurationTime, because the method already does it